refactor(docs-infra): extract `NodeRuntimeSandbox` lazy-loading into separate function
In this commit, we extract `NodeRuntimeSandbox` lazy-loading into a separate function (because it's duplicated).
