+++ adev/src/app/editor/index.ts
@@ -9,8 +9,7 @@
 export {EmbeddedTutorialManager} from './embedded-tutorial-manager.service';
 export {LoadingStep} from './enums/loading-steps';
 export {NodeRuntimeState} from './node-runtime-state.service';
-
-export {NodeRuntimeSandbox} from './node-runtime-sandbox.service';
+export {injectNodeRuntimeSandbox} from './inject-node-runtime-sandbox';
 
 export {EmbeddedEditor, EMBEDDED_EDITOR_SELECTOR} from './embedded-editor.component';
